+++ ShootingStar/src/application/MainMenu.java
@@ -88,6 +88,7 @@ public class MainMenu extends Application {
         primaryStage.setScene(scene);
         primaryStage.setTitle("Shooting Star [Main Menu][alpha]");
         primaryStage.setResizable(false);
+        primaryStage.setAlwaysOnTop(true);
         primaryStage.show();
     }