+++ ShootingStar/src/player/Star.java
@@ -3,8 +3,11 @@ package player;
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Circle;
 import javafx.scene.input.KeyCode;
+
 import java.util.HashSet;
 import java.util.Set;
+
+import enemies.Enemy;
 
 
 public class Star {
@@ -15,6 +18,7 @@ public class Star {
     private Movement movement;
     private double sceneWidth;
     private double sceneHeight;
+    private boolean collided;
     
     private static final double OBJECT_RADIUS = 20.0;
     private Set<KeyCode> pressedKeys = new HashSet<>();   
@@ -26,11 +30,13 @@ public class Star {
     	this.movement = new Movement();
     	this.sceneWidth = sceneWidth;
     	this.sceneHeight = sceneHeight;
+    	this.collided = false;
     }
 
     public Circle getObject() {
     	return object;
     }
+  
     
     public void handleKeyPress(KeyCode keyCode) {
     	pressedKeys.add(keyCode);
@@ -60,24 +66,42 @@ public class Star {
         }
     }
     
+    
+    public void handleCollisions(Enemy enemy) {
+    	if (enemy.isCollidedWithStar(this)) {
+            collided = true;
+            enemy.hasCollided(true);
+            stopMovement(); // Stop Star's movement upon collision
+    	}
+	}
+    
+	public boolean isCollided() {
+		return collided;
+	}
+    
     public void updatePosition() {
-    	objectX += movement.getVelocityX();
-    	objectY += movement.getVelocityY() + Movement.WIND_FORCE;
-
-    	// Boundary checking
-    	double minX = OBJECT_RADIUS;
-        double minY = -100; // Allows the star to go beyond the top of the window
-        double maxX = sceneWidth - OBJECT_RADIUS;
-        double maxY = sceneHeight - OBJECT_RADIUS;
-
-        if (objectX < minX) objectX = minX;
-        if (objectY < minY) objectY = minY;
-        if (objectX > maxX) objectX = maxX;
-        if (objectY > maxY) objectY = maxY;
-        
-        // Update position of the object
-        object.setCenterX(objectX);
-    	object.setCenterY(objectY);
+    	if (!collided) {
+	    	objectX += movement.getVelocityX();
+	    	objectY += movement.getVelocityY() + Movement.WIND_FORCE;
+	
+	    	// Boundary checking
+	    	double minX = OBJECT_RADIUS;
+	        double minY = -100; // Allows the star to go beyond the top of the window
+	        double maxX = sceneWidth - OBJECT_RADIUS;
+	        double maxY = sceneHeight - OBJECT_RADIUS;
+	
+	        if (objectX < minX) objectX = minX;
+	        if (objectY < minY) objectY = minY;
+	        if (objectX > maxX) objectX = maxX;
+	        if (objectY > maxY) objectY = maxY;
+	        
+	        // Update position of the object
+	        object.setCenterX(objectX);
+	    	object.setCenterY(objectY);
+    	} else {
+    		movement.stopVertical();
+    		movement.stopHorizontal();
+    	}
     }
     
 	public boolean isOutOfBounds() {