+++ ShootingStar/src/screens/GameScreen.java
@@ -8,10 +8,11 @@ import javafx.scene.text.Font;
 import javafx.scene.text.FontWeight;
 import javafx.scene.text.Text;
 import javafx.scene.text.TextAlignment;
-import javafx.util.Duration;
-
 import player.Star;
+
+import enemies.Enemy;
 import enemies.Rocket;
+import enemies.Bird;
 
 public class GameScreen {
 	private Group root;
@@ -19,8 +20,17 @@ public class GameScreen {
     private Text messageText;
     private Text timerText;
     private Text gameOverText;
+    private Text generalTimerText;
     private double outOfBoundsTimer = 5.0;
-    private Rocket rocket;
+    private double generalTimer = 0.0;
+    private boolean abort = false;
+    
+    private int enemySpawnCount = 0;
+    private final int MAX_ENEMY_COUNT = 6;
+    private final long NANOSECONDS_PER_SECOND = 1_000_000_000;
+    private long lastSpawnTime = 0;
+    private final long spawnInterval = NANOSECONDS_PER_SECOND; // Spawn an enemy every second
+
 	
 	public GameScreen() {
 		root = new Group();
@@ -29,12 +39,13 @@ public class GameScreen {
 		star = new Star(scene.getWidth(), scene.getHeight());		
 		root.getChildren().add(star.getObject());
 		
-		rocket = new Rocket(scene.getWidth(), scene.getHeight());
-		root.getChildren().add(rocket.getObject());
-		
 		gameOverText = createGameOverText();
 		root.getChildren().add(gameOverText);
 		gameOverText.setVisible(false);
+		
+		generalTimerText = createGeneralTimerText();
+		root.getChildren().add(generalTimerText);
+		generalTimerText.setVisible(true);
 		
 		messageText = createMessageText();
 		timerText = createTimerText();
@@ -45,11 +56,21 @@ public class GameScreen {
 		AnimationTimer timer = new AnimationTimer() {
             @Override
             public void handle(long now) {
-            	rocket.updatePosition();
+            	long currentTime = System.nanoTime();
+            	
+            	if (currentTime - lastSpawnTime >= spawnInterval && enemySpawnCount < MAX_ENEMY_COUNT) {
+                    spawnEnemy();
+                    lastSpawnTime = currentTime;
+                }
+            	
+            	for (Enemy enemy: Enemy.getEnemies()) {
+            		enemy.updatePosition();
+            		star.handleCollisions(enemy);
+            	}
                 star.updatePosition();
                 handleCollisions();
-                rocket.updateSurvivalTime(Duration.seconds(rocket.getSurvivalTime().toSeconds()));
                 handleOutOfBoundsMessages();
+                handleTimer();
             }
         };
         timer.start(); // Start the timer to continuously update the star's position
@@ -59,17 +80,29 @@ public class GameScreen {
 		return root.getScene();
 	}
 	
+	private void spawnEnemy() {
+        // Create and add a new rocket to the scene
+		Enemy rocket = new Rocket(getScene().getWidth(), getScene().getHeight());
+        root.getChildren().add(rocket.getObject());
+        Enemy.addEnemy(rocket);
+        
+        Enemy bird = new Bird(getScene().getWidth(), getScene().getHeight());
+        root.getChildren().add(bird.getObject());
+        Enemy.addEnemy(bird);
+        enemySpawnCount++;
+    }
+	
 	private void handleCollisions() { 
-		if (rocket.isCollidedWithStar(star)){
+		
+		if (star.isCollided()){
 			stopGame();
 		}
 	}
 	
 	private void stopGame() {
 		star.stopMovement();
+		abort = true;
 		
-		gameOverText.setText("Game Over! You survived for: " + (int) rocket.getSurvivalTime().toSeconds() + " seconds");
-		gameOverText.setVisible(true);
 	}
 	
 	private Text createGameOverText() {
@@ -80,6 +113,23 @@ public class GameScreen {
         text.setLayoutX(root.getScene().getWidth() / 2 - 200);
         text.setLayoutY(50);
         return text;
+	}
+	
+	private void handleTimer() {
+		if (!abort) generalTimer += 0.016;
+		else {
+			generalTimerText.setFont(Font.font("Arial", FontWeight.BOLD, 30));
+	    	generalTimerText.setTextAlignment(TextAlignment.CENTER);
+	    	generalTimerText.setLayoutX((root.getScene().getWidth() - generalTimerText.getLayoutBounds().getWidth()) / 2);
+	    	generalTimerText.setLayoutY(root.getScene().getHeight() / 2);
+		}
+		
+		if(!root.getChildren().contains(generalTimerText)) {
+			root.getChildren().add(generalTimerText);
+		}
+		
+		generalTimerText.setText(String.format("Time elapsed: %.2f", generalTimer));
+		
 	}
 	
     private void handleOutOfBoundsMessages() {
@@ -124,5 +174,18 @@ public class GameScreen {
         text.setLayoutY(100);
         text.setVisible(false);
         return text;
+    }
+    
+    private Text createGeneralTimerText() {
+    	Text text = new Text("Time elapsed: ");
+    	text.setFont(Font.font("Arial", FontWeight.BOLD, 18));
+        text.setFill(Color.BLACK);
+        text.setTextAlignment(TextAlignment.RIGHT);
+        text.setLayoutX(root.getScene().getWidth() - 50 - text.getLayoutBounds().getWidth());
+        text.setLayoutY(30);
+        text.setVisible(true);
+        text.toFront();
+        return text;
+    	
     }
 }