+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.0.3 (2023-12-7 13:03)
+
+### Release Highlights
+- Added two types of enemies: Bird and Rocket
+- Feature: Elapsed time implemented
+
 # 0.0.2 (2023-12-6 16:09)
 
 ### Release Highlights
@@ -10,7 +16,7 @@
 - Feature: Added movement and controls to player object
 - Feature: Implemented wind drag
 - Feature: Bounding boxes added
-- Feature: Timer implemented for a message to  pop up when the player has been out of bounds (top) for too long.
+- Feature: Timer implemented for a message to pop up when the player has been out of bounds (top) for too long.
 
 # 0.0.0 (2023-11-28 0-7:50)